requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
trustcall
langgraph-cli[inmem] 
trustcall
langchain_experimental
langgraph-supervisor

langchain-mongodb
pymongo[srv]    
motor

notebook
tavily-python

langchain_anthropic
langchain_openai

pydantic
fastapi[standard]
websockets
uvicorn


beautifulsoup4
nest_asyncio
pypdf
pytest
pytest-asyncio
pytest-mock
pytest-cov

langchain-groq